# src/main/java/com/deemor/motif/helpRequest/dto/HelpRequestAddDto.java
package com.deemor.motif.helpRequest.dto;

import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
public class HelpRequestAddDto {
    private String description;
}